crm_kazakhstan/requirements.txt
django>=3.2,<4.0
python-decouple==3.8
celery[redis]==5.3.4
django-celery-beat==2.5.0
psycopg2-binary==2.9.9
flower==1.2.0
freezegun==0.3.4
django-timezone-field==6.1.0
django-guardian==2.4.0
django-admin-rangefilter==0.12.0
django-admin-interface==0.28.5
aiogram==3.4.1